feat(batch-11): S3 versioning — 383 passing (27 new versioning tests)
- Versioning state machine: put/get ?versioning per bucket
- put_object_v2 returns (version_id, old_extents) — GC inside metadata layer
- delete_specific_version + get_object_version with versionId query param
- ListObjectVersions: <Version>/<DeleteMarker> elements, pagination, common prefixes
- DeleteObjects dispatches to delete_specific_version when VersionId provided
- ACL stubs: get_object_acl, get_bucket_acl, put_acl (valid XML, no enforcement)
- Dual-lookup for "null" versionId: resolves both version_id='' and version_id='null'
- Suspended versioning: also removes version_id='' rows when overwriting null-version
- x-amz-version-id suppressed for suspended/disabled (only emitted for real UUIDs)
- CompleteMultipartUpload: removed pre-emptive delete_metadata that wiped version history
- Split handlers.rs (3237 lines) into handlers/ module with 11 focused files
