+++ server/src/metadata/sqlite_store.rs
@@ -3,6 +3,7 @@
 use crate::metadata::{MetadataStorage, Metadata, ObjectId, BucketStats};
 use crate::util::serializer::serialize_offset_size;
 use std::sync::Mutex;
+use std::sync::atomic::{AtomicU64, Ordering};
 use rusqlite::{params, Connection};
 use std::sync::Arc;
 use log::{warn, info, error};
@@ -10,6 +11,40 @@ use actix_web::Error;
 use lazy_static::lazy_static;
 use std::env;
 use std::path::{Path, PathBuf};
+
+static VERSION_COUNTER: AtomicU64 = AtomicU64::new(0);
+
+fn generate_version_id() -> String {
+    let counter = VERSION_COUNTER.fetch_add(1, Ordering::SeqCst);
+    let ts = chrono::Utc::now().timestamp_nanos_opt().unwrap_or(0) as u64;
+    format!("{:016x}{:016x}", ts, counter)
+}
+
+/// Result of a versioning-aware delete (no explicit versionId given).
+pub enum VersioningDeleteResult {
+    /// Versioning disabled — object data removed (or never existed).
+    Deleted,
+    /// Versioning enabled/suspended — a delete marker was created.
+    Marker { version_id: String },
+}
+
+/// Result of permanently deleting a specific version.
+pub struct DeleteSpecificResult {
+    pub found: bool,
+    pub was_delete_marker: bool,
+    pub version_id: String,
+}
+
+/// One entry in a ListObjectVersions response.
+pub struct VersionRow {
+    pub key: String,
+    pub version_id: String,
+    pub is_delete_marker: bool,
+    pub etag: String,
+    pub size: u64,
+    pub last_modified: String,
+    pub is_latest: bool,
+}
 
 fn get_db_path() -> PathBuf {
     match env::var("DB_FILE") {
@@ -36,13 +71,19 @@ lazy_static! {
 
         conn.execute_batch("PRAGMA journal_mode=WAL;").ok();
 
-        // Object metadata table — stores S3 object extents + S3 metadata fields
+        // Object metadata table — one row per (user, bucket, key, version_id).
+        // version_id='' means versioning is disabled for that bucket.
+        // version_id='null' means the null-version in a suspended-versioning bucket.
+        // is_latest=1 marks the current visible version for a key.
         conn.execute(
             "CREATE TABLE IF NOT EXISTS objects (
                 id               INTEGER PRIMARY KEY AUTOINCREMENT,
                 user             TEXT NOT NULL,
                 bucket           TEXT NOT NULL,
                 key              TEXT NOT NULL,
+                version_id       TEXT NOT NULL DEFAULT '',
+                is_delete_marker INTEGER NOT NULL DEFAULT 0,
+                is_latest        INTEGER NOT NULL DEFAULT 1,
                 offset_size_list BLOB,
                 etag             TEXT,
                 size             INTEGER NOT NULL DEFAULT 0,
@@ -52,16 +93,11 @@ lazy_static! {
                 cache_control    TEXT,
                 expires          TEXT,
                 content_encoding TEXT,
-                UNIQUE(user, bucket, key)
+                parts_manifest   TEXT,
+                UNIQUE(user, bucket, key, version_id)
             )",
             [],
         ).expect("Failed to create objects table");
-
-        // Migrate existing databases that predate these columns
-        conn.execute("ALTER TABLE objects ADD COLUMN cache_control TEXT", []).ok();
-        conn.execute("ALTER TABLE objects ADD COLUMN expires TEXT", []).ok();
-        conn.execute("ALTER TABLE objects ADD COLUMN content_encoding TEXT", []).ok();
-        conn.execute("ALTER TABLE objects ADD COLUMN parts_manifest TEXT", []).ok();
 
         // Multipart upload tracking tables
         conn.execute_batch(
@@ -103,9 +139,10 @@ lazy_static! {
         // Bucket registry — tracks created buckets (including empty ones)
         conn.execute(
             "CREATE TABLE IF NOT EXISTS buckets (
-                user       TEXT NOT NULL,
-                name       TEXT NOT NULL,
-                created_at TEXT NOT NULL DEFAULT (strftime('%Y-%m-%dT%H:%M:%S.000Z', 'now')),
+                user              TEXT NOT NULL,
+                name              TEXT NOT NULL,
+                created_at        TEXT NOT NULL DEFAULT (strftime('%Y-%m-%dT%H:%M:%S.000Z', 'now')),
+                versioning_state  TEXT NOT NULL DEFAULT 'disabled',
                 PRIMARY KEY (user, name)
             )",
             [],
@@ -160,16 +197,28 @@ impl SQLiteMetadataStore {
 
 impl MetadataStorage for SQLiteMetadataStore {
     fn put_metadata(&self, user_id: &str, bucket: &str, object_id: &str, metadata: &Metadata) -> Result<(), Error> {
+        // Non-versioned write: DELETE existing version_id='' row then INSERT new one.
         let offset_size_list = metadata.to_offset_size_list();
         let offset_size_bytes = serialize_offset_size(&offset_size_list)?;
         let user_metadata_json = serde_json::to_string(&metadata.user_metadata)
             .unwrap_or_else(|_| "{}".to_string());
 
         let conn = DB_CONN.lock().unwrap();
+        conn.execute(
+            "DELETE FROM objects WHERE user = ?1 AND bucket = ?2 AND key = ?3 AND version_id = ''",
+            params![user_id, bucket, object_id],
+        ).map_err(actix_web::error::ErrorInternalServerError)?;
+        // Also clear is_latest on all existing versions so the non-versioned row becomes latest.
+        conn.execute(
+            "UPDATE objects SET is_latest = 0 WHERE user = ?1 AND bucket = ?2 AND key = ?3 AND is_latest = 1",
+            params![user_id, bucket, object_id],
+        ).map_err(actix_web::error::ErrorInternalServerError)?;
         let result = conn.execute(
             "INSERT INTO objects
-                (user, bucket, key, offset_size_list, etag, size, content_type, last_modified, user_metadata, cache_control, expires, content_encoding)
-             VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9, ?10, ?11, ?12)",
+                (user, bucket, key, version_id, is_latest, is_delete_marker,
+                 offset_size_list, etag, size, content_type, last_modified,
+                 user_metadata, cache_control, expires, content_encoding)
+             VALUES (?1, ?2, ?3, '', 1, 0, ?4, ?5, ?6, ?7, ?8, ?9, ?10, ?11, ?12)",
             params![
                 user_id, bucket, object_id,
                 offset_size_bytes,
@@ -195,13 +244,15 @@ impl MetadataStorage for SQLiteMetadataStore {
     fn get_metadata(&self, user_id: &str, bucket: &str, object_id: &str) -> Result<Metadata, Error> {
         let conn = DB_CONN.lock().unwrap();
         let mut stmt = conn.prepare(
-            "SELECT offset_size_list, etag, size, content_type, last_modified, user_metadata, cache_control, expires, content_encoding
-             FROM objects WHERE user = ?1 AND bucket = ?2 AND key = ?3",
+            "SELECT offset_size_list, etag, size, content_type, last_modified, user_metadata,
+                    cache_control, expires, content_encoding, version_id, is_delete_marker
+             FROM objects
+             WHERE user = ?1 AND bucket = ?2 AND key = ?3 AND is_latest = 1",
         ).map_err(actix_web::error::ErrorInternalServerError)?;
 
         let row = stmt.query_row(params![user_id, bucket, object_id], |row| {
             Ok((
-                row.get::<_, Vec<u8>>(0)?,
+                row.get::<_, Option<Vec<u8>>>(0)?,
                 row.get::<_, Option<String>>(1)?,
                 row.get::<_, i64>(2)?,
                 row.get::<_, Option<String>>(3)?,
@@ -210,6 +261,8 @@ impl MetadataStorage for SQLiteMetadataStore {
                 row.get::<_, Option<String>>(6)?,
                 row.get::<_, Option<String>>(7)?,
                 row.get::<_, Option<String>>(8)?,
+                row.get::<_, String>(9)?,
+                row.get::<_, i64>(10)?,
             ))
         }).map_err(|e| {
             warn!("get_metadata: not found user={} bucket={} key={}: {}", user_id, bucket, object_id, e);
@@ -218,8 +271,20 @@ impl MetadataStorage for SQLiteMetadataStore {
             ))
         })?;
 
-        let (offset_size_bytes, etag, size, content_type, last_modified, user_metadata_json, cache_control, expires, content_encoding) = row;
-        let offset_size_list = crate::util::serializer::deserialize_offset_size(&offset_size_bytes)?;
+        let (offset_size_bytes, etag, size, content_type, last_modified, user_metadata_json,
+             cache_control, expires, content_encoding, version_id, is_delete_marker) = row;
+
+        if is_delete_marker != 0 {
+            return Err(actix_web::error::ErrorNotFound(format!(
+                "No data found for key: {} in bucket: {}, The key does not exist", object_id, bucket
+            )));
+        }
+
+        let offset_size_list = if let Some(bytes) = offset_size_bytes {
+            crate::util::serializer::deserialize_offset_size(&bytes)?
+        } else {
+            vec![]
+        };
         let user_metadata: std::collections::HashMap<String, String> = user_metadata_json
             .as_deref()
             .and_then(|s| serde_json::from_str(s).ok())
@@ -234,10 +299,12 @@ impl MetadataStorage for SQLiteMetadataStore {
         metadata.cache_control = cache_control;
         metadata.expires = expires;
         metadata.content_encoding = content_encoding;
+        metadata.version_id = if version_id.is_empty() { None } else { Some(version_id) };
         Ok(metadata)
     }
 
     fn delete_metadata(&self, user_id: &str, bucket: &str, object_id: &str) -> Result<(), Error> {
+        // Hard-delete all rows for this key (used by CompleteMultipartUpload overwrite and internal cleanup).
         let conn = DB_CONN.lock().unwrap();
         conn.execute(
             "DELETE FROM objects WHERE user = ?1 AND bucket = ?2 AND key = ?3",
@@ -249,7 +316,9 @@ impl MetadataStorage for SQLiteMetadataStore {
     fn list_objects(&self, user_id: &str, bucket: &str) -> Result<Vec<ObjectId>, Error> {
         let conn = DB_CONN.lock().unwrap();
         let mut stmt = conn.prepare(
-            "SELECT key FROM objects WHERE user = ?1 AND bucket = ?2 ORDER BY key",
+            "SELECT key FROM objects
+             WHERE user = ?1 AND bucket = ?2 AND is_latest = 1 AND is_delete_marker = 0
+             ORDER BY key",
         ).map_err(actix_web::error::ErrorInternalServerError)?;
 
         let rows = stmt.query_map(params![user_id, bucket], |row| {
@@ -265,11 +334,12 @@ impl MetadataStorage for SQLiteMetadataStore {
 
     fn object_exists(&self, user_id: &str, bucket: &str, object_id: &str) -> Result<bool, Error> {
         let conn = DB_CONN.lock().unwrap();
-        let mut stmt = conn.prepare(
-            "SELECT COUNT(*) FROM objects WHERE user = ?1 AND bucket = ?2 AND key = ?3",
+        let count: i64 = conn.query_row(
+            "SELECT COUNT(*) FROM objects
+             WHERE user = ?1 AND bucket = ?2 AND key = ?3 AND is_latest = 1 AND is_delete_marker = 0",
+            params![user_id, bucket, object_id],
+            |row| row.get(0),
         ).map_err(actix_web::error::ErrorInternalServerError)?;
-        let count: i64 = stmt.query_row(params![user_id, bucket, object_id], |row| row.get(0))
-            .map_err(actix_web::error::ErrorInternalServerError)?;
         Ok(count > 0)
     }
 
@@ -291,7 +361,7 @@ impl MetadataStorage for SQLiteMetadataStore {
                 cache_control    = ?7,
                 expires          = ?8,
                 content_encoding = ?9
-             WHERE user = ?10 AND bucket = ?11 AND key = ?12",
+             WHERE user = ?10 AND bucket = ?11 AND key = ?12 AND is_latest = 1",
             params![
                 offset_size_bytes,
                 metadata.etag, metadata.size as i64,
@@ -328,6 +398,7 @@ impl MetadataStorage for SQLiteMetadataStore {
                     COALESCE(SUM(o.size), 0) AS total_size
              FROM   buckets b
              LEFT JOIN objects o ON o.user = b.user AND o.bucket = b.name
+                                 AND o.is_latest = 1 AND o.is_delete_marker = 0
              WHERE  b.user = ?1
              GROUP BY b.name
              ORDER BY b.name",
@@ -394,7 +465,8 @@ impl MetadataStorage for SQLiteMetadataStore {
     fn bucket_object_stats(&self, user_id: &str, bucket: &str) -> Result<(u64, u64), Error> {
         let conn = DB_CONN.lock().unwrap();
         let mut stmt = conn.prepare(
-            "SELECT COUNT(*), COALESCE(SUM(size), 0) FROM objects WHERE user = ?1 AND bucket = ?2",
+            "SELECT COUNT(*), COALESCE(SUM(size), 0) FROM objects
+             WHERE user = ?1 AND bucket = ?2 AND is_latest = 1 AND is_delete_marker = 0",
         ).map_err(actix_web::error::ErrorInternalServerError)?;
         let (count, bytes): (i64, i64) = stmt.query_row(params![user_id, bucket], |row| {
             Ok((row.get(0)?, row.get(1)?))
@@ -627,6 +699,369 @@ impl SQLiteMetadataStore {
             Err(rusqlite::Error::QueryReturnedNoRows) => Ok(String::new()),
             Err(e) => Err(actix_web::error::ErrorInternalServerError(e)),
         }
+    }
+}
+
+/// Versioning operations
+impl SQLiteMetadataStore {
+    pub fn get_versioning_state(&self, bucket: &str) -> Result<String, Error> {
+        let conn = DB_CONN.lock().unwrap();
+        let result: rusqlite::Result<Option<String>> = conn.query_row(
+            "SELECT versioning_state FROM buckets WHERE name = ?1",
+            params![bucket],
+            |row| row.get(0),
+        );
+        match result {
+            Ok(s) => Ok(s.unwrap_or_else(|| "disabled".to_string())),
+            Err(rusqlite::Error::QueryReturnedNoRows) => Ok("disabled".to_string()),
+            Err(e) => Err(actix_web::error::ErrorInternalServerError(e)),
+        }
+    }
+
+    pub fn set_versioning_state(&self, user_id: &str, bucket: &str, state: &str) -> Result<(), Error> {
+        let conn = DB_CONN.lock().unwrap();
+        conn.execute(
+            "UPDATE buckets SET versioning_state = ?1 WHERE user = ?2 AND name = ?3",
+            params![state, user_id, bucket],
+        ).map_err(actix_web::error::ErrorInternalServerError)?;
+        Ok(())
+    }
+
+    /// Versioning-aware PUT. Returns Some(version_id) when versioning is enabled/suspended,
+    /// None when versioning is disabled (overwrites in place).
+    /// Write a new object version.
+    /// Returns (version_id, old_extents_to_gc).
+    /// - version_id: Some(vid) when versioning enabled/suspended, None when disabled.
+    /// - old_extents_to_gc: extents of the row that was replaced (caller should queue for GC).
+    pub fn put_object_v2(
+        &self, user_id: &str, bucket: &str, key: &str, metadata: &Metadata,
+    ) -> Result<(Option<String>, Vec<(u64, u64)>), Error> {
+        let versioning = self.get_versioning_state(bucket)?;
+        let offset_size_list = metadata.to_offset_size_list();
+        let offset_size_bytes = serialize_offset_size(&offset_size_list)?;
+        let user_metadata_json = serde_json::to_string(&metadata.user_metadata)
+            .unwrap_or_else(|_| "{}".to_string());
+
+        let conn = DB_CONN.lock().unwrap();
+
+        match versioning.as_str() {
+            "disabled" => {
+                // Read old extents before deleting so the caller can GC them.
+                let old_extents: Vec<(u64, u64)> = conn.query_row(
+                    "SELECT offset_size_list FROM objects WHERE user=?1 AND bucket=?2 AND key=?3 AND version_id=''",
+                    params![user_id, bucket, key],
+                    |row| row.get::<_, Option<Vec<u8>>>(0),
+                ).unwrap_or(None)
+                .and_then(|b| crate::util::serializer::deserialize_offset_size(&b).ok())
+                .unwrap_or_default();
+
+                conn.execute(
+                    "DELETE FROM objects WHERE user=?1 AND bucket=?2 AND key=?3 AND version_id=''",
+                    params![user_id, bucket, key],
+                ).map_err(actix_web::error::ErrorInternalServerError)?;
+                conn.execute(
+                    "INSERT INTO objects
+                        (user,bucket,key,version_id,is_latest,is_delete_marker,
+                         offset_size_list,etag,size,content_type,last_modified,
+                         user_metadata,cache_control,expires,content_encoding,parts_manifest)
+                     VALUES(?1,?2,?3,'',1,0,?4,?5,?6,?7,?8,?9,?10,?11,?12,?13)",
+                    params![user_id,bucket,key,offset_size_bytes,metadata.etag,
+                            metadata.size as i64,metadata.content_type,metadata.last_modified,
+                            user_metadata_json,metadata.cache_control,metadata.expires,
+                            metadata.content_encoding,metadata.properties.get("parts_manifest")],
+                ).map_err(actix_web::error::ErrorInternalServerError)?;
+                Ok((None, old_extents))
+            }
+            "enabled" => {
+                let vid = generate_version_id();
+                // Demote current latest; old versions stay in DB (versioning preserves them).
+                conn.execute(
+                    "UPDATE objects SET is_latest=0 WHERE user=?1 AND bucket=?2 AND key=?3 AND is_latest=1",
+                    params![user_id, bucket, key],
+                ).map_err(actix_web::error::ErrorInternalServerError)?;
+                conn.execute(
+                    "INSERT INTO objects
+                        (user,bucket,key,version_id,is_latest,is_delete_marker,
+                         offset_size_list,etag,size,content_type,last_modified,
+                         user_metadata,cache_control,expires,content_encoding,parts_manifest)
+                     VALUES(?1,?2,?3,?4,1,0,?5,?6,?7,?8,?9,?10,?11,?12,?13,?14)",
+                    params![user_id,bucket,key,vid,offset_size_bytes,metadata.etag,
+                            metadata.size as i64,metadata.content_type,metadata.last_modified,
+                            user_metadata_json,metadata.cache_control,metadata.expires,
+                            metadata.content_encoding,metadata.properties.get("parts_manifest")],
+                ).map_err(actix_web::error::ErrorInternalServerError)?;
+                Ok((Some(vid), vec![]))
+            }
+            _ /* "suspended" */ => {
+                // Read extents from whichever null-variant exists: 'null' (prior suspended write)
+                // or '' (object written before versioning was ever enabled).
+                let extents_from = |vid: &str| -> Vec<(u64, u64)> {
+                    conn.query_row(
+                        "SELECT offset_size_list FROM objects WHERE user=?1 AND bucket=?2 AND key=?3 AND version_id=?4",
+                        params![user_id, bucket, key, vid],
+                        |row| row.get::<_, Option<Vec<u8>>>(0),
+                    ).unwrap_or(None)
+                    .and_then(|b| crate::util::serializer::deserialize_offset_size(&b).ok())
+                    .unwrap_or_default()
+                };
+                let old_extents: Vec<(u64, u64)> = {
+                    let ne = extents_from("null");
+                    if !ne.is_empty() { ne } else { extents_from("") }
+                };
+
+                // Delete both the null-version and the pre-versioning non-versioned row.
+                conn.execute(
+                    "DELETE FROM objects WHERE user=?1 AND bucket=?2 AND key=?3 AND version_id='null'",
+                    params![user_id, bucket, key],
+                ).map_err(actix_web::error::ErrorInternalServerError)?;
+                conn.execute(
+                    "DELETE FROM objects WHERE user=?1 AND bucket=?2 AND key=?3 AND version_id=''",
+                    params![user_id, bucket, key],
+                ).map_err(actix_web::error::ErrorInternalServerError)?;
+                conn.execute(
+                    "UPDATE objects SET is_latest=0 WHERE user=?1 AND bucket=?2 AND key=?3 AND is_latest=1",
+                    params![user_id, bucket, key],
+                ).map_err(actix_web::error::ErrorInternalServerError)?;
+                conn.execute(
+                    "INSERT INTO objects
+                        (user,bucket,key,version_id,is_latest,is_delete_marker,
+                         offset_size_list,etag,size,content_type,last_modified,
+                         user_metadata,cache_control,expires,content_encoding,parts_manifest)
+                     VALUES(?1,?2,?3,'null',1,0,?4,?5,?6,?7,?8,?9,?10,?11,?12,?13)",
+                    params![user_id,bucket,key,offset_size_bytes,metadata.etag,
+                            metadata.size as i64,metadata.content_type,metadata.last_modified,
+                            user_metadata_json,metadata.cache_control,metadata.expires,
+                            metadata.content_encoding,metadata.properties.get("parts_manifest")],
+                ).map_err(actix_web::error::ErrorInternalServerError)?;
+                Ok((Some("null".to_string()), old_extents))
+            }
+        }
+    }
+
+    /// Versioning-aware DELETE (no explicit versionId).
+    pub fn delete_object_v2(&self, user_id: &str, bucket: &str, key: &str) -> Result<VersioningDeleteResult, Error> {
+        let versioning = self.get_versioning_state(bucket)?;
+        let conn = DB_CONN.lock().unwrap();
+
+        match versioning.as_str() {
+            "disabled" => {
+                conn.execute(
+                    "DELETE FROM objects WHERE user=?1 AND bucket=?2 AND key=?3",
+                    params![user_id, bucket, key],
+                ).map_err(actix_web::error::ErrorInternalServerError)?;
+                Ok(VersioningDeleteResult::Deleted)
+            }
+            "enabled" => {
+                let vid = generate_version_id();
+                let now = chrono::Utc::now().format("%Y-%m-%dT%H:%M:%S.000Z").to_string();
+                conn.execute(
+                    "UPDATE objects SET is_latest=0 WHERE user=?1 AND bucket=?2 AND key=?3 AND is_latest=1",
+                    params![user_id, bucket, key],
+                ).map_err(actix_web::error::ErrorInternalServerError)?;
+                conn.execute(
+                    "INSERT INTO objects
+                        (user,bucket,key,version_id,is_latest,is_delete_marker,size,last_modified)
+                     VALUES(?1,?2,?3,?4,1,1,0,?5)",
+                    params![user_id, bucket, key, vid, now],
+                ).map_err(actix_web::error::ErrorInternalServerError)?;
+                Ok(VersioningDeleteResult::Marker { version_id: vid })
+            }
+            _ /* "suspended" */ => {
+                // Delete the existing null-version object (if it exists) and insert a null-version delete marker.
+                conn.execute(
+                    "DELETE FROM objects WHERE user=?1 AND bucket=?2 AND key=?3 AND version_id='null'",
+                    params![user_id, bucket, key],
+                ).map_err(actix_web::error::ErrorInternalServerError)?;
+                conn.execute(
+                    "UPDATE objects SET is_latest=0 WHERE user=?1 AND bucket=?2 AND key=?3 AND is_latest=1",
+                    params![user_id, bucket, key],
+                ).map_err(actix_web::error::ErrorInternalServerError)?;
+                let now = chrono::Utc::now().format("%Y-%m-%dT%H:%M:%S.000Z").to_string();
+                conn.execute(
+                    "INSERT INTO objects
+                        (user,bucket,key,version_id,is_latest,is_delete_marker,size,last_modified)
+                     VALUES(?1,?2,?3,'null',1,1,0,?4)",
+                    params![user_id, bucket, key, now],
+                ).map_err(actix_web::error::ErrorInternalServerError)?;
+                Ok(VersioningDeleteResult::Marker { version_id: "null".to_string() })
+            }
+        }
+    }
+
+    /// Permanently delete a specific version (DELETE ?versionId=x).
+    /// Returns info about what was deleted.
+    pub fn delete_specific_version(&self, user_id: &str, bucket: &str, key: &str, version_id: &str) -> Result<DeleteSpecificResult, Error> {
+        let conn = DB_CONN.lock().unwrap();
+
+        // "null" from the client matches both '' (versioning never enabled) and 'null' (suspended null-version).
+        let effective_vid: &str = if version_id == "null" {
+            let exists_empty: bool = conn.query_row(
+                "SELECT 1 FROM objects WHERE user=?1 AND bucket=?2 AND key=?3 AND version_id=''",
+                params![user_id, bucket, key], |_| Ok(true),
+            ).unwrap_or(false);
+            if exists_empty { "" } else { "null" }
+        } else {
+            version_id
+        };
+
+        // Fetch the row first to know is_delete_marker and is_latest.
+        let row_result: rusqlite::Result<(i64, i64)> = conn.query_row(
+            "SELECT is_delete_marker, is_latest FROM objects WHERE user=?1 AND bucket=?2 AND key=?3 AND version_id=?4",
+            params![user_id, bucket, key, effective_vid],
+            |row| Ok((row.get(0)?, row.get(1)?)),
+        );
+
+        match row_result {
+            Err(rusqlite::Error::QueryReturnedNoRows) => {
+                return Ok(DeleteSpecificResult { found: false, was_delete_marker: false, version_id: version_id.to_string() });
+            }
+            Err(e) => return Err(actix_web::error::ErrorInternalServerError(e)),
+            Ok(_) => {}
+        }
+        let (is_dm, was_latest) = row_result.unwrap();
+
+        // Delete the row.
+        conn.execute(
+            "DELETE FROM objects WHERE user=?1 AND bucket=?2 AND key=?3 AND version_id=?4",
+            params![user_id, bucket, key, effective_vid],
+        ).map_err(actix_web::error::ErrorInternalServerError)?;
+
+        // If the deleted row was is_latest, promote the most-recently-inserted remaining row.
+        if was_latest != 0 {
+            conn.execute(
+                "UPDATE objects SET is_latest=1 WHERE id=(
+                     SELECT id FROM objects WHERE user=?1 AND bucket=?2 AND key=?3 ORDER BY id DESC LIMIT 1
+                 )",
+                params![user_id, bucket, key],
+            ).map_err(actix_web::error::ErrorInternalServerError)?;
+        }
+
+        Ok(DeleteSpecificResult {
+            found: true,
+            was_delete_marker: is_dm != 0,
+            version_id: version_id.to_string(),
+        })
+    }
+
+    /// Fetch a specific version of an object.
+    pub fn get_object_version(&self, user_id: &str, bucket: &str, key: &str, version_id: &str) -> Result<Metadata, Error> {
+        let conn = DB_CONN.lock().unwrap();
+        // "null" matches '' (versioning never enabled) or 'null' (suspended null-version).
+        let effective_vid: &str = if version_id == "null" {
+            let exists_empty: bool = conn.query_row(
+                "SELECT 1 FROM objects WHERE user=?1 AND bucket=?2 AND key=?3 AND version_id=''",
+                params![user_id, bucket, key], |_| Ok(true),
+            ).unwrap_or(false);
+            if exists_empty { "" } else { "null" }
+        } else {
+            version_id
+        };
+        let row = conn.query_row(
+            "SELECT offset_size_list,etag,size,content_type,last_modified,user_metadata,
+                    cache_control,expires,content_encoding,version_id,is_delete_marker
+             FROM objects WHERE user=?1 AND bucket=?2 AND key=?3 AND version_id=?4",
+            params![user_id, bucket, key, effective_vid],
+            |row| Ok((
+                row.get::<_, Option<Vec<u8>>>(0)?,
+                row.get::<_, Option<String>>(1)?,
+                row.get::<_, i64>(2)?,
+                row.get::<_, Option<String>>(3)?,
+                row.get::<_, Option<String>>(4)?,
+                row.get::<_, Option<String>>(5)?,
+                row.get::<_, Option<String>>(6)?,
+                row.get::<_, Option<String>>(7)?,
+                row.get::<_, Option<String>>(8)?,
+                row.get::<_, String>(9)?,
+                row.get::<_, i64>(10)?,
+            )),
+        ).map_err(|e| {
+            if e == rusqlite::Error::QueryReturnedNoRows {
+                actix_web::error::ErrorNotFound("Version not found")
+            } else {
+                actix_web::error::ErrorInternalServerError(e)
+            }
+        })?;
+
+        let (offset_size_bytes, etag, size, content_type, last_modified, user_metadata_json,
+             cache_control, expires, content_encoding, vid, is_delete_marker) = row;
+
+        let offset_size_list = if let Some(bytes) = offset_size_bytes {
+            crate::util::serializer::deserialize_offset_size(&bytes)?
+        } else {
+            vec![]
+        };
+        let user_metadata = user_metadata_json
+            .as_deref()
+            .and_then(|s| serde_json::from_str(s).ok())
+            .unwrap_or_default();
+
+        let mut metadata = Metadata::from_offset_size_list(offset_size_list);
+        metadata.etag = etag;
+        metadata.size = size as u64;
+        metadata.content_type = content_type;
+        metadata.last_modified = last_modified;
+        metadata.user_metadata = user_metadata;
+        metadata.cache_control = cache_control;
+        metadata.expires = expires;
+        metadata.content_encoding = content_encoding;
+        metadata.version_id = if vid.is_empty() { None } else { Some(vid) };
+        metadata.is_delete_marker = is_delete_marker != 0;
+        Ok(metadata)
+    }
+
+    /// List all versions for a bucket (for GET ?versions).
+    /// Returns (rows, is_truncated, next_key_marker, next_version_id_marker).
+    pub fn list_object_versions_full(
+        &self, user_id: &str, bucket: &str,
+        prefix: &str, key_marker: &str, version_id_marker: &str, max_keys: usize,
+    ) -> Result<(Vec<VersionRow>, bool, String, String), Error> {
+        let conn = DB_CONN.lock().unwrap();
+        let mut stmt = conn.prepare(
+            "SELECT key, version_id, is_delete_marker, etag, size, last_modified, is_latest
+             FROM objects
+             WHERE user=?1 AND bucket=?2
+             ORDER BY key ASC, id DESC",
+        ).map_err(actix_web::error::ErrorInternalServerError)?;
+
+        let rows = stmt.query_map(params![user_id, bucket], |row| {
+            Ok(VersionRow {
+                key:               row.get(0)?,
+                version_id:        row.get(1)?,
+                is_delete_marker:  row.get::<_, i64>(2)? != 0,
+                etag:              row.get::<_, Option<String>>(3)?.unwrap_or_default(),
+                size:              row.get::<_, i64>(4)? as u64,
+                last_modified:     row.get::<_, Option<String>>(5)?.unwrap_or_default(),
+                is_latest:         row.get::<_, i64>(6)? != 0,
+            })
+        }).map_err(actix_web::error::ErrorInternalServerError)?;
+
+        let mut all: Vec<VersionRow> = rows
+            .filter_map(|r| r.ok())
+            .filter(|r| r.key.starts_with(prefix))
+            .collect();
+
+        // Apply key-marker / version-id-marker pagination.
+        if !key_marker.is_empty() {
+            let mut past_marker = false;
+            all.retain(|r| {
+                if past_marker { return true; }
+                if r.key.as_str() > key_marker { past_marker = true; return true; }
+                if r.key.as_str() == key_marker && !version_id_marker.is_empty() && r.version_id.as_str() > version_id_marker {
+                    return true;
+                }
+                false
+            });
+        }
+
+        let is_truncated = all.len() > max_keys;
+        let page: Vec<VersionRow> = all.into_iter().take(max_keys).collect();
+        let (next_key, next_vid) = if is_truncated {
+            page.last().map(|r| (r.key.clone(), r.version_id.clone())).unwrap_or_default()
+        } else {
+            (String::new(), String::new())
+        };
+
+        Ok((page, is_truncated, next_key, next_vid))
     }
 }
